+++ src/azure-cli/azure/cli/command_modules/monitor/tests/latest/test_monitor_general_operations.py
@@ -146,50 +146,6 @@ class MonitorClonePublicIpScenarios(ScenarioTest):
             ])
 
 
-class MonitorClonePublicIpAlwaysScenarios(ScenarioTest):
-    @ResourceGroupPreparer(name_prefix='cli_test_metric_alert_clone')
-    def test_monitor_clone_public_ip_metric_alerts_always_scenario(self, resource_group):
-        self.test_guid_count = 0
-        self.kwargs.update({
-            'alert': 'alert1',
-            'alert2': 'alert2',
-            'ag1': 'ag1',
-            'sub': self.get_subscription_id(),
-            'rg': resource_group,
-            'ip1': 'ip1',
-            'ip2': 'ip2'
-        })
-
-        ip1_json = self.cmd('network public-ip create -g {rg} -n {ip1}').get_output_in_json()
-        ip2_json = self.cmd('network public-ip create -g {rg} -n {ip2}').get_output_in_json()
-        self.kwargs.update({
-            'ip1_id': ip1_json['publicIp']['id'],
-            'ip2_id': ip2_json['publicIp']['id']
-        })
-
-        self.cmd('monitor action-group create -g {rg} -n {ag1}')
-        self.cmd('monitor metrics alert create -g {rg} -n {alert} --scopes {ip1_id} --action {ag1} --description "Test" --condition "total TCPBytesForwardedDDoS > 5"', checks=[
-            self.check('description', 'Test'),
-            self.check('severity', 2),
-            self.check('windowSize', '0:05:00'),
-            self.check('evaluationFrequency', '0:01:00'),
-            self.check('length(scopes)', 1)
-        ])
-
-        self.cmd('monitor metrics alert create -g {rg} -n {alert2} --scopes {ip1_id} --action {ag1} --description "Test2" --condition "max TCPBytesForwardedDDoS > 5"', checks=[
-            self.check('description', 'Test2'),
-            self.check('severity', 2),
-            self.check('windowSize', '0:05:00'),
-            self.check('evaluationFrequency', '0:01:00'),
-            self.check('length(scopes)', 1)
-        ])
-
-        with mock.patch('azure.cli.command_modules.monitor.util._gen_guid', side_effect=self.create_guid):
-            self.cmd('monitor clone --source-resource {ip1_id} --target-resource {ip2_id}', checks=[
-                self.check('length(metricsAlert)', 2),
-            ])
-
-
 class MonitorCloneStorageAccountAcrossSubsScenarios(ScenarioTest):
     @live_only()
     @ResourceGroupPreparer(name_prefix='cli_test_metric_alert_clone')
